tweaks

+++ experimental/utils/timer.h
@@ -2,7 +2,7 @@
 class timer {
 public:
     chrono::time_point<chrono::high_resolution_clock> startTime, endTime;
-    chrono::duration<float> duration;
+    chrono::duration<double> duration;
 
     void start() {
         startTime = chrono::high_resolution_clock::now();
@@ -12,7 +12,7 @@ public:
         endTime = chrono::high_resolution_clock::now();
         duration = endTime - startTime;
 
-        //time in milliseconds
+        // Time in milliseconds
         cout << "Runtime: " << duration.count() * 1000 << "ms\n";
     }
-}t;
+} t;
